fix(auth):dont lose context when reloading
