+++ src/App.css
@@ -1,62 +1,24 @@
 #root {
-  max-width: 1280px;
-  text-align: center;
-}
-
-:root{
-  --cor01: #0B0A07;
-  --cor02: #1F1F1F;
-  --cor03: #D6D6D6;
-  --cor04: #70C3FF;
-}
-
-
-*{
+  all: unset;
+  overflow-x: hidden;
+  height: 100vh;
+  width: 100%;
+  max-width: 100%;
+  background-color: #f4f2ee;
+  border: 0;
   margin: 0;
   padding: 0;
-  box-sizing: border-box;
-
-}
-
-body{
-  width: 100vw;
-  height: 100vh;
-  background-color: #f0f2f5;
   font-family: Arial, Helvetica, sans-serif
 }
 
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
+.container {
+  -ms-overflow-style: none;  /* Internet Explorer 10+ */
+  scrollbar-width: none;  /* Firefox */
 }
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
+.container::-webkit-scrollbar { 
+  display: none;  /* Safari and Chrome */
 }
 
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  a:nth-of-type(2) .logo {
-    animation: logo-spin infinite 20s linear;
-  }
-}
-
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
-}
+/* * { 
+  outline: 1px solid red; 
+} */
