+++ src/pages/Home/HomeLayout.tsx
@@ -1,38 +1,32 @@
-import { Box } from '@mui/material'
-// import Sidebar from '../../components/Sidebar'
 import styled from 'styled-components'
 import { Outlet } from 'react-router-dom'
+import { Box } from '@mui/material'
 import Header from './components/Header'
-import Left from './components/Sidebar'
+import Right from './components/Right'
+import Left from './components/Left'
 
 const Layout = styled.div`
-  display: grid;
-  grid-template-areas: 'left main right';
-  grid-template-columns: minmax(0, 5fr) minmax(0, 12fr) minmax(300px, 7fr);
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  justify-content: space-between;
   column-gap: 25px;
-  row-gap: 25px;
   padding: 0px 32px;
+  width: 100%;
+  max-width: 96.68%;
   margin: 25px 0;
-  @media (max-width: 768px) {
-    display: flex;
-    flex-direction: column;
-    padding: 0 5px;
-  }
 `
 
 function HomeLayout() {
   return (
-    <>
+    <Box sx={{ backgroundColor: '#f4f2ee', height: '100vh', overflow: 'hidden' }}>
       <Header />
       <Layout>
-        <>
-          <Box />
-          <Box />
-        </>
-        <Outlet />
         <Left />
+        <Outlet />
+        <Right />
       </Layout>
-    </>
+    </Box>
   )
 }
