+++ src/pages/Home/components/CompetencyCard.tsx
@@ -1,18 +1,12 @@
 import styled from 'styled-components';
-import { Container, Typography } from '@mui/material';
+import { Box, Typography } from '@mui/material';
 import { CompetenciaLevel } from '../../../services/endpoints/competencia';
-
-
-
-const CardTitle = styled(Typography)`
-  margin-bottom: 16px;
-  font-weight: bold;
-`;
 
 const CompetencyContainer = styled.div`
   display: flex;
   align-items: center;
-  margin-bottom: 12px;
+  justify-content: flex-start;
+  padding: 8px;
 `;
 
 const CompetencyName = styled(Typography)`
@@ -20,25 +14,30 @@ const CompetencyName = styled(Typography)`
 `;
 
 const LevelIcon = styled.img`
-  width: 40px;
-  height: 40px;
+  width: 32px;
+  height: 32px;
+  padding: 0px 8px;
 `;
 
 interface CompetencyCardProps {
-    competencies: CompetenciaLevel[];
+  competencies: CompetenciaLevel[];
 }
 
 function CompetencyCard({ competencies }: CompetencyCardProps) {
   return (
-    <Container sx={{ height: "300px", borderRadius: "8px", overflowY: "scroll", overflowX: "hidden" }}>
-      <CardTitle variant="h6">Competências</CardTitle>
-      {competencies.map((comp, index) => (
-        <CompetencyContainer key={index}>
-          <CompetencyName variant="body1">{comp.competencia.descricao}</CompetencyName>
-          <LevelIcon src={`/images/level-${comp.level}.svg`} alt={`Nível ${comp.level}`} />
-        </CompetencyContainer>
-      ))}
-    </Container>
+    <Box className="container" sx={{ height: '300px', borderRadius: '8px', padding: '0px 16px', overflowY: 'auto', overflowX: 'hidden', display: 'flex', justifyContent: 'flex-start', alignItems: 'center', flexDirection: 'column', scrollbarGutter: 'stable' }}>
+      <Box sx={{ padding: '16px', position: 'sticky', top: 0, backgroundColor: 'white', zIndex: 1 }}>
+        <Typography variant="span" sx={{ fontSize: '16px', fontWeight: 600, whiteSpace: 'nowrap' }}>Competências quentes</Typography>
+      </Box>
+      <Box sx={{ display: 'flex', justifyContent: 'flex-start', alignItems: 'flex-start', flexDirection: 'column', width: '100%' }}>
+        {competencies.map((comp, index) => (
+          <CompetencyContainer key={index}>
+            <LevelIcon src={`/images/level-${comp.level}.svg`} alt={`Nível ${comp.level}`} />
+            <CompetencyName variant="body1">{comp.competencia.descricao}</CompetencyName>
+          </CompetencyContainer>
+        ))}
+      </Box>
+    </Box>
   );
 }
