+++ src/pages/Home/components/Right.tsx
@@ -1,0 +1,76 @@
+import React from 'react'
+import styled from 'styled-components'
+
+const Container = styled.div`
+  min-width: 200px;
+`
+
+const ArtCard = styled.div`
+  text-align: center;
+  overflow: hidden;
+  margin-bottom: 8px;
+  border-radius: 5px;
+  background-color: #fff;
+  transition: box-shadow 83ms;
+  position: relative;
+  border: none;
+  box-shadow:
+    0 0 0 1px rgb(0 0 0 / 15%),
+    0 0 0 rgb(0 0 0 / 20%);
+`
+
+const UserInfo = styled.div`
+  border-bottom: 1px solid rgba(0, 0, 0, 0.15);
+  padding: 12px 12px 16px;
+  word-wrap: break-word;
+  word-break: break-word;
+`
+
+const CardBackground = styled.div`
+  background: url('/images/card-bg.svg');
+  background-position: center;
+  background-size: 462px;
+  height: 54px;
+  margin: -12px -12px 0;
+`
+
+const Photo = styled.div`
+  box-shadow: none;
+  background: url(${(props) => props.photoUrl});
+  width: 72px;
+  height: 72px;
+  box-sizing: border-box;
+  background-clip: content-box;
+  background-color: #fff;
+  background-position: center;
+  background-size: 150%;
+  background-repeat: no-repeat;
+  border: 2px solid white;
+  margin: -38px auto 12px;
+  border-radius: 50%;
+`
+
+const Text = styled.div`
+  font-size: 16px;
+  line-height: 1.5;
+  color: rgba(0, 0, 0, 0.9);
+  font-weight: 600;
+`
+
+function Right() {
+  return (
+    <Container>
+      <ArtCard>
+        <UserInfo>
+          <CardBackground />
+          <Photo photoUrl="https://pbs.twimg.com/media/FTkEZfxWIAA0f4P.jpg" />
+          <Text>
+            Welcome
+          </Text>
+        </UserInfo>
+      </ArtCard>
+    </Container>
+  )
+}
+
+export default Right
