+++ src/services/endpoints/_axios.ts
@@ -33,6 +33,19 @@ _axios.interceptors.response.use(
   }
 )
 
+_axios.interceptors.request.use(
+  config => {
+    const token = localStorage.getItem('token')
+    if (token) {
+      config.headers['Authorization'] = `Bearer ${token}`
+    }
+    return config
+  },
+  error => {
+    return Promise.reject(error)
+  }
+)
+
 export type AxiosResponse<T> = Partial<Omit<_AxiosResponse, 'data'>> & {
   data: {data: T }
 }
