+++ src/styles/global.tsx
@@ -12,7 +12,6 @@ const GlobalStyle = createGlobalStyle`
 body{
     width: 100vw;
     height: 100vh;
-    background-color: #f0f2f5;
     font-family: Arial, Helvetica, sans-serif
 }
